Disable spellcheck for HTML5 compatible clients
Improve the performance of the webpage app by turning off the spellchecking API for that particular textarea element in browsers for significant performance gains on slower computers and / or working with large blocks of code.  The spellchecking feature of browsers is ineffective without a programming language dictionary and programming language patterns.